% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theta.R, R/cadlag.R
\name{cadlag.fun}
\alias{cadlag.fun}
\title{Cadlag linear functions}
\usage{
cadlag.fun(x, y, order = 0)

cadlag.fun(x, y, order = 0)
}
\arguments{
\item{x}{c() of x values}

\item{y}{c() of y values of length 1 greater than x}

\item{order}{0 for constant and 1 for linear}

\item{x}{c() of x values}

\item{y}{c() of y values of length 1 greater than x}

\item{order}{0 for constant and 1 for linear}
}
\description{
This function allows you to express your love of cats.

This function allows you to express your love of cats.
}
\examples{
cadlag.fun(c(0,1),c(1,2,3))
cadlag.fun(c(0,1),c(1,2),order=1)

cadlag.fun(c(0,1),c(1,2,3))
cadlag.fun(c(0,1),c(1,2),order=1)

}
\keyword{cadlag}
