% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theta.R, R/cadlag.R
\name{cadlag}
\alias{cadlag}
\title{Cadlag linear functions}
\usage{
cadlag(x, y = FALSE, order = 0)

cadlag(x, y = FALSE, order = 0)
}
\arguments{
\item{x}{c() of x values}

\item{y}{c() of y values of length 1 greater than x}

\item{x}{c() of x values}

\item{y}{c() of y values of length 1 greater than x}
}
\description{
This function allows you to express your love of cats.

This function allows you to express your love of cats.
}
\examples{
cadlag(c(0,1),c(1,2,3))

cadlag(c(0,1),c(1,2,3))

}
\keyword{cadlag}
