% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theta.R, R/cadlag.R
\name{integrate.cadlag}
\alias{integrate.cadlag}
\title{Indefinite integral of cadlag functions}
\usage{
integrate.cadlag(f, lower = NA, upper = NA, ..., subdivisions = 100L,
  rel.tol = .Machine$double.eps^0.25, abs.tol = rel.tol,
  stop.on.error = TRUE, keep.xy = FALSE, aux = NULL)

integrate.cadlag(f, lower = NA, upper = NA, ..., subdivisions = 100L,
  rel.tol = .Machine$double.eps^0.25, abs.tol = rel.tol,
  stop.on.error = TRUE, keep.xy = FALSE, aux = NULL)
}
\arguments{
\item{f}{object of type cadlag}

\item{f}{object of type cadlag}
}
\value{
object of type cadlag

object of type cadlag
}
\description{
This function allows you to express your love of cats.

This function allows you to express your love of cats.
}
\examples{
f = cadlag(c(0,1),c(1,2,3))
F = integrate.cadlag(f)

f = cadlag(c(0,1),c(1,2,3))
F = integrate.cadlag(f)

}
\keyword{cadlag}
