% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theta.R
\name{Theta_hat}
\alias{Theta_hat}
\title{Compute Theta_hat statistic for a dataset}
\usage{
Theta_hat(surv_data)
}
\arguments{
\item{surv_data}{A data.frame (see the example)}
}
\value{
An object with the statistic
}
\description{
This function allows you to express your love of cats.
}
\examples{
require(survival)
surv_data = with(subset(survival::lung,ph.ecog \%in\% 0:2),
data.frame(population = sex,
           censor = as.numeric(status==1),
           time = time, cohort = ph.ecog ))
out = Theta_hat(surv_data)
print(out)
print(confint(out))
print(pvalue.Theta_hat(out))



}
